+++ src/Extensions/OptionSpanExtensions.cs
@@ -6,7 +6,8 @@ namespace Ametrin.Optional;
 public static class OptionSpanExtensions
 {
     public static Option<T> Parse<T>(this RefOption<ReadOnlySpan<char>> option, IFormatProvider? provider = null) where T : ISpanParsable<T>
-        => option._hasValue && T.TryParse(option._value, provider, out var result) ? Option.Success(result) : default;
+        => option._hasValue ? OptionParsingExtensions.TryParse<T>(option._value, provider) : default;
+
     public static RefOption<Span<T>> RejectEmpty<T>(this RefOption<Span<T>> option)
         => option.Reject(static span => span.IsEmpty);
     public static RefOption<ReadOnlySpan<T>> RejectEmpty<T>(this RefOption<ReadOnlySpan<T>> option)